fix: apply code review fixes across firmware sources
- Remove _codeql_detected_source_root symlink (CodeQL artifact)
- .gitignore: remove Ninja project entries, add SDCC/CMake artifact patterns, add _codeql_detected_source_root
- copilot-instructions.md: add 256B IRAM to memory constraints line
- board.c: fix backslash includes to forward slashes, remove unused stdlib.h/stdio.h
- board.c: remove no-op ENHIT |= 0x00, clarify IT0 falling-edge comment
- board.c: add explicit 4-bit masking for n_DAT[0] bit-packing
- board.c: fix TOCTOU in check_zc - clear zero_crossed immediately on detection
- board.c: remove dead unreachable if-block in calculateSkip
- board.h: fix swapped heater/NTC pin comments for coffee (P1.6/PWM2) and steam (P1.7/PWM3)
- ADC.c: clear ADCIF flag in ISR, remove large commented-out dead code block
- ADC.h: remove unimplemented ADC_Init and ADC declarations
- IIC.c: add bounds checks on n_Addr before all n_DAT accesses
- timer.c: make halTick volatile, make getTick atomic via ET0 disable/enable
- pid.c: widen derivative to long to prevent 16-bit overflow
- sensors.c: fix backslash include to forward slash
- sensors.h: add #include "board.h" to make header self-contained
- I2C.md: remove "generated by Grok" trailing attribution line

Agent-Logs-Url: https://github.com/banoz/CRM3007L.FW/sessions/394d1610-cb91-40eb-b68e-4894ed15e05a

+++ src/pid.c
@@ -79,7 +79,7 @@ void pid_reset(void)
 unsigned char pid_tick(unsigned int current_temp, unsigned int setpoint)
 {
 	int error;
-	int derivative;
+	long derivative;
 	long output;
 	long output_sum;
 	long p_term;
@@ -109,7 +109,7 @@ unsigned char pid_tick(unsigned int current_temp, unsigned int setpoint)
 	}
 	error = (int)setpoint - (int)current_temp;
 
-	derivative = error - pid_last_error;
+	derivative = (long)error - (long)pid_last_error;
 	pid_last_error = error;
 
 	pid_read_gains(&kp, &ki, &kd);
